+++ openTEPES/__init__.py
@@ -2,8 +2,8 @@
 
 
     Args:
-        case: Name of the folder where the CSV files of the case are found
-        dir: Main path where the case folder can be found
+        case:   Name of the folder where the CSV files of the case are found
+        dir:    Main path where the case folder can be found
         solver: Name of the solver
 
     Returns:
@@ -14,10 +14,10 @@
         >>> oT.routine("9n", "C:\\Users\\UserName\\Documents\\GitHub\\openTEPES", "glpk")
 """
 
-__version__ = '2.2.0'
-from .openTEPES_Main import main
-from .openTEPES import *
-from .openTEPES_InputData import *
+__version__ = '2.2.1'
+from .openTEPES_Main             import main
+from .openTEPES                  import *
+from .openTEPES_InputData        import *
 from .openTEPES_ModelFormulation import *
-from .openTEPES_OutputResults import *
-from .openTEPES_ProblemSolving import *
+from .openTEPES_OutputResults    import *
+from .openTEPES_ProblemSolving   import *