+++ CHANGELOG.rst
@@ -4,10 +4,14 @@ Change Log
 - [CHANGED] Addition of the sequence of decisions and operation (TEP + transmission switching)
 - [FIXED] fixing `<https://pascua.iit.comillas.edu/aramos/openTEPES/index.html#>`_ in README.rst 
 
+[2.2.1] - 2021-04-02
+----------------------
+- [CHANGED] added three new output files for line commitment, switch on and off
+
 [2.2.0] - 2021-03-31
 ----------------------
 - [CHANGED] introduction of Power-to-X in ESS. Modifies the Generation file and introduces a new EnergyOutflows file
-- [CHANGED] introduction of switching decision for transmission lines
+- [CHANGED] introduction of switching decision for transmission lines. Modifies the Option file and introduces a new column BinarySwitching in Network file
 
 [2.1.0] - 2021-03-18
 ----------------------